intermediate commit

+++ src/main/resources/v_1_0__init.sql
@@ -7,5 +7,5 @@ create table "table"
     id           serial primary key,
     name         varchar(255) unique,
     participants int not null default 0,
-    ordering     int not null default nextval('ordering_num_seq')
+    ordering     bigint not null default nextval('ordering_num_seq')
 );
